refactor: CartAppBar를 CommonAppBar로 대체 및 컴포넌트 패키지 이동
