+++ app/src/main/java/woowacourse/shopping/feature/productlist/component/ProductItem.kt
@@ -35,7 +35,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import woowacourse.shopping.R
-import woowacourse.shopping.feature.common.ProductQuantitySelector
+import woowacourse.shopping.feature.common.component.ProductQuantitySelector
 
 @Composable
 fun ProductItem(
